First approach of screens

+++ feature-country/src/main/java/com/damazo/featurecountry/model/Coordinate.kt
@@ -1,6 +1,6 @@
 package com.damazo.featurecountry.model
 
 data class Coordinate(
-    val longitude: Long,
-    val latitude: Long,
+    val longitude: Double,
+    val latitude: Double,
 )
